@@ -9,48 +9,48 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@aws-sdk/client-ses": "^3.282.0",
-    "@headlessui/react": "^1.7.11",
+    "@aws-sdk/client-ses": "^3.301.0",
+    "@headlessui/react": "^1.7.13",
     "@headwayapp/react-widget": "^0.0.4",
     "@heroicons/react": "^2.0.16",
-    "@lottiefiles/react-lottie-player": "^3.5.2",
+    "@lottiefiles/react-lottie-player": "^3.5.3",
     "@next-auth/prisma-adapter": "^1.0.5",
-    "@next/font": "^13.1.6",
-    "@prisma/client": "^4.10.1",
-    "@react-email/components": "0.0.2",
+    "@next/font": "^13.2.4",
+    "@prisma/client": "^4.12.0",
+    "@react-email/components": "0.0.4",
     "@react-email/render": "0.0.6",
-    "@react-email/tailwind": "0.0.6",
-    "@sentry/nextjs": "^7.38.0",
-    "@stripe/stripe-js": "^1.47.0",
+    "@react-email/tailwind": "0.0.7",
+    "@sentry/nextjs": "^7.45.0",
+    "@stripe/stripe-js": "^1.52.0",
     "axios": "^1.3.4",
-    "crisp-sdk-web": "^1.0.13",
-    "firebase": "^9.17.1",
-    "immer": "^9.0.19",
+    "crisp-sdk-web": "^1.0.15",
+    "firebase": "^9.18.0",
+    "immer": "^9.0.21",
     "iron-session": "^6.3.1",
-    "jotai": "^2.0.2",
+    "jotai": "^2.0.3",
     "jotai-immer": "^0.2.0",
-    "next": "^13.1.6",
-    "next-auth": "^4.19.2",
+    "next": "^13.2.4",
+    "next-auth": "^4.20.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-frame-component": "^5.2.6",
-    "recharts": "^2.4.3",
-    "stripe": "^11.11.0",
-    "swr": "^2.0.3"
+    "recharts": "^2.5.0",
+    "stripe": "^11.16.0",
+    "swr": "^2.1.1"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.3",
-    "@types/node": "^18.14.1",
-    "@types/react": "^18.0.28",
+    "@types/node": "^18.15.11",
+    "@types/react": "^18.0.31",
     "@types/react-dom": "^18.0.11",
-    "autoprefixer": "^10.4.13",
-    "eslint": "^8.34.0",
-    "eslint-config-next": "^13.1.6",
+    "autoprefixer": "^10.4.14",
+    "eslint": "^8.37.0",
+    "eslint-config-next": "^13.2.4",
     "postcss": "^8.4.21",
-    "prettier": "^2.8.4",
-    "prettier-plugin-tailwindcss": "^0.2.3",
-    "prisma": "^4.10.1",
-    "tailwindcss": "^3.2.7",
-    "typescript": "^4.9.5"
+    "prettier": "^2.8.7",
+    "prettier-plugin-tailwindcss": "^0.2.5",
+    "prisma": "^4.12.0",
+    "tailwindcss": "^3.3.0",
+    "typescript": "^5.0.2"
   }
 }
